projet final GAI
You have been employed as a Data Scientist by a consultancy firm. The firm has a client who is a used car dealer. They have a special feature on Ford cars and they want your firm to design a model that can predict the optimum quotation price for the cars in their lot. They provide you with sales data for the past few years. The dataset contains different features of the cars and the price they were sold at.

The tasks assigned to you are as follows.

There might be a few duplicate entries and a few missing values in the dataset. Data cleaning will be a part of the assignment.
You have to perform exploratory data analysis to draw keen insights on the data and determine the effect of different features on the price. Some specific requests by the client include:
a. Identify number of sales for each fuel type
b. Identify which transmission type has more price outliers
Compare the models with linear, polynomial and ridge regressions on single and multiple variables to find the best performing model
Perform a Grid Search on the Ridge regression model to identify the optimum hyperparameter for the model for best performance.
You decide to use Generative AI to create python codes that can help you analyse the data, determine the best features and create the prediction model as per requirement.